chore: code refactor for clarity
